% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alignedExample}
\alias{alignedExample}
\title{Two protein sequences xyz coordinates after alignment.}
\format{Raw data to be used and filtered in 3D visualization.}
\usage{
alignedExample
}
\description{
A pdb file for two sequences after alignment. Used in alignment3dVisualization
 function
}
\keyword{datasets}
