% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment3dVisualization.R
\name{alignment3dVisualization}
\alias{alignment3dVisualization}
\title{alignment3dVisualization}
\usage{
alignment3dVisualization(alignedPath, targetName, templateName)
}
\arguments{
\item{alignedPath}{The filepath of pdb after alignment.}

\item{targetName}{A string for the pdb name for the first part of
the pdb file after alignment.}

\item{templateName}{A string for the pdb name for the second part of
the pdb file after alignment.}
}
\value{
return nothing but a window for visualization of two proteins in 3D
}
\description{
Visualization of how similar two sequences are in 3D. Note that
 the input pdb file must be the pdb after alignment. The blue one
 should be the template and the red one should be the target. A recommend
 webpage is http://fatcat.burnham.org/fatcat-cgi/cgi/fatcat.pl?-func=pairwise
 Download the pdb file after submitting the alignemnt as the input as this function.
}
\examples{
# http://fatcat.burnham.org/fatcat-cgi/cgi/fatcat.pl?-func=pairwise
# submit two files to the link and dowload the pdb file, call
# it align.pdb

alignment3dVisualization("./inst/extdata/alignedExample.pdb", "4GSA", "5TE2")

}
