% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{databaseExample}
\alias{databaseExample}
\title{A pdb file.}
\format{Raw data to be used and filtered in selectTemplates function.}
\usage{
databaseExample
}
\description{
A fasta file about the information of ipr005814 family whose sequences are
reviewed and have 3D structure. Sequences and entry names are included.
}
\keyword{datasets}
