% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{databaseExampleText}
\alias{databaseExampleText}
\title{Sequence database pdb information.}
\format{Raw data to be used in selectTemplates function.}
\usage{
databaseExampleText
}
\description{
A text file about the information of ipr005814 family whose sequences are
reviewed and have corresponding pdb. Pdb names are included
}
\keyword{datasets}
