% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPdbFileGenerator.R
\name{GetPdbFileGenerator}
\alias{GetPdbFileGenerator}
\title{GetPdbFileGenerator}
\usage{
GetPdbFileGenerator()
}
\value{
return nothing but output required pdb files
}
\description{
Get pdb file from RCSB website. To run this function
induvisually, ensure there is a pdbList.txt in current
work directory containing the pdb name. pdbList.txt
will be deleted after the execution.
}
\examples{
cat("4GSA \n 4A0R \n", file = "pdbList.txt")
GetPdbFileGenerator()

}
