% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathGeneratorMovement.R
\name{pathGeneratorMovement}
\alias{pathGeneratorMovement}
\title{selectTemplates}
\usage{
pathGeneratorMovement(fileName, suffix = NULL, destination = "",
  move = FALSE)
}
\arguments{
\item{fileName}{File name(s) in a character vector. It must in
the current work directory. Make sure file(s) exists.}

\item{suffix}{(optional) A character string representingthe
suffix of the file, or NULL. Default is NULL.}

\item{destination}{(optional) The directory name if not doing
file moving, or the directory path if doing file moving.
Default is "". Make sure the destination exists.}

\item{move}{(optional) A boolean. TRUE for file moving, FALSE
for generating file path. Default is FALSE.}
}
\value{
\itemize{
\item If move = TRUE, return nothing, but the input file(s)
 are moved to the destination
\item If move = FALSE, return generated path(s)
}
}
\description{
File movements in disk
}
\examples{
# move myFile.txt to C:/Users
cat("test", file = "myFile.txt")
pathGeneratorMovement(fileName = "myFile", suffix = ".txt",
                      destination = "C:/Users/new", move = TRUE)
pathGeneratorMovement(fileName = "myFile.txt",
                      destination = "C:/Users/new", move = TRUE)

# generate file path "C:/Users/new/myFile.txt".
pathGeneratorMovement(fileName = "myFile", suffix = ".txt",
                      destination = "/new", move = FALSE)
pathGeneratorMovement(fileName = "myFile.txt",
                      destination = "/new")

}
