% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runHmtp.R
\name{runHmtp}
\alias{runHmtp}
\title{3dAlignment}
\usage{
runHmtp()
}
\value{
No return value but open up a shiny page
}
\description{
Launch the shiny app for package gmtp
A function that launches the shiny app for this package.
The code has been placed in \code{./inst/shiny-scripts}
}
\examples{
\donyrun{
runHmtp()
}

}
