% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{targetExample}
\alias{targetExample}
\title{Target Sequence information.}
\format{Raw data to be used in selectTemplates function.}
\usage{
targetExample
}
\description{
A fasta file about the information of A3SNF5. Sequences and entry names are
 included.
}
\keyword{datasets}
